+++ src/beignet/_periodic_displacement.py
@@ -2,25 +2,23 @@ import torch
 from torch import Tensor
 
 
-def periodic_displacement(box: float | Tensor, dR: Tensor) -> Tensor:
+def periodic_displacement(input: Tensor, position: Tensor) -> Tensor:
     r"""Wraps displacement vectors into a hypercube.
 
-    Parameters
+    Parameters:
     ----------
     box : float or Tensor
         Specification of hypercube size. Either:
-        (a) float if all sides have equal length.
+        (a) scalar if all sides have equal length.
         (b) Tensor of shape (spatial_dim,) if sides have different lengths.
+
     dR : Tensor
         Matrix of displacements with shape (..., spatial_dim).
 
-    Returns
+    Returns:
     -------
-    Tensor
+    output : Tensor, shape=(...)
         Matrix of wrapped displacements with shape (..., spatial_dim).
     """
-    distances = (
-        torch.remainder(dR + box * torch.tensor(0.5, dtype=torch.float32), box)
-        - torch.tensor(0.5, dtype=torch.float32) * box
-    )
-    return distances
+    output = torch.remainder(position + input * 0.5, input) - 0.5 * input
+    return output
